Move back button to component
